googlecloudplatform/requirements.txt
tensorflow==2.13.1
google-cloud-storage
Pillow
numpy